requirements.txt
langchain>=0.2.0
langchain-anthropic>=0.1.15
langchain-community>=0.2.0
langchain-core>=0.2.0
tavily-python>=0.3.0
wikipedia>=1.4.0
python-dotenv>=1.0.0
anthropic>=0.25.0
